@@ -26,9 +26,9 @@
     "@sightread/flake": "^0.5.6",
     "clsx": "^1.1.1",
     "jasmid.ts": "^2.1.0",
-    "next": "^12.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.1",
+    "next": "^12.1.6",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "typescript": "^4.3.5"
   },
   "scripts": {
@@ -38,7 +38,7 @@
     "lint": "next lint",
     "format": "prettier --write src/**/*.ts*",
     "gen": "cd scripts && ts-node generate.ts && cd ..",
-    "render": "RENDER=1 ts-node --project scripts/tsconfig.json scripts/render.ts",
+    "render": "RENDER=1 ts-node --project scripts/tsconfig.json scripts/render.ts --huge-max-old-generation-size",
     "record": "cd scripts && ts-node screenrecorder.ts && cd ..",
     "storybook": "start-storybook -p 6006"
   },
