@@ -1,691 +1,16 @@
-export type MusicFile = { type: 'song'; file: string; artist: string; difficulty: string }
+import { MusicFile } from '@/types'
 
-const songs: MusicFile[] = [
-  {
-    file: 'Fur_Elise.mid',
-    name: 'Fur Elise',
-    artist: 'Beethoven',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Abba - Dancing Queen.mid',
-    name: 'Dancing Queen',
-    artist: 'Abba',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Alan Walker - Darkside.mid',
-    name: 'Darkside',
-    artist: 'Alan Walker',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Alan Walker - On My Way.mid',
-    name: 'On My Way',
-    artist: 'Alan Walker',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Alan Walker - The Spectre.mid',
-    name: 'The Spectre',
-    artist: 'Alan Walker',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Alan_Walker_Faded.mid',
-    name: 'Faded',
-    artist: 'Alan Walker',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Alec Benjamin - Let Me Down Slowly.mid',
-    name: 'Let Me Down Slowly',
-    artist: 'Alec Benjamin',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Ariana Grande - Only 1.mid',
-    name: 'Only 1',
-    artist: 'Ariana Grande',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Attack_On_Titan_Theme_Guren_No_Yumiya.mid',
-    name: 'Attack On Titan Theme Song',
-    artist: 'Hiroyuki Sawano',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'BTS - Black Swan.mid',
-    name: 'Black Swan',
-    artist: 'BTS',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'BTS - Spring Day.mid',
-    name: 'Spring Day',
-    artist: 'BTS',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Backstreet Boys - I Need You Tonight.mid',
-    name: 'I Need You Tonight',
-    artist: 'Backstreet Boys',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Billie Eilish - Bad Guy.mid',
-    name: 'Bad Guy',
-    artist: 'Billie Eilish',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Billie Eilish - Six Feet Under.mid',
-    name: 'Six Feet Under',
-    artist: 'Billie Eilish',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Billie Eilish ft. Khalid - Lovely.mid',
-    name: 'Lovely',
-    artist: 'Billie Eilish ft. Khalid',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Breathe_No_More_-_Evanescence.mid',
-    name: 'Breathe No More',
-    artist: 'Evanescence',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Bruno Mars - Count On Me.mid',
-    name: 'Count On Me',
-    artist: 'Bruno Mars',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Bruno Mars - Talking to the Moon.mid',
-    name: 'Talking to the Moon',
-    artist: 'Bruno Mars',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Calum Scott - Dancing On My Own.mid',
-    name: 'Dancing On My Own',
-    artist: 'Calum Scott',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Calum Scott - You Are The Reason.mid',
-    name: 'You Are The Reason',
-    artist: 'Calum Scott',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Canon_Rock.xml',
-    name: 'Canon Rock',
-    artist: 'JerryC',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Clair_de_Lune.mid',
-    name: 'Clair de Lune',
-    artist: 'Isao Tomita',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Coldplay - Clocks.mid',
-    name: 'Clocks',
-    artist: 'Coldplay',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Craig Armstrong - Love Actually - Soundtrack.mid',
-    name: 'Love Actually',
-    artist: 'Craig Armstrong',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Curb Your Enthusiasm.mid',
-    name: 'Curb Your Enthusiasm Theme Song',
-    artist: 'Luciano Michelini',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Demi Lovato - Cool For The Summer.mid',
-    name: 'Cool For The Summer',
-    artist: 'Demi Lovato',
-    difficulty: 'Easy',
-  },
-  {
-    file: "Dua Lipa - Don't Start Now.mid",
-    name: "Don't Start Now",
-    artist: 'Dua Lipa',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Ed Sheeran - Nancy Mulligan.mid',
-    name: 'Nancy Mulligan',
-    artist: 'Ed Sheeran',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Ed Sheeran - Perfect.mid',
-    name: 'Perfect',
-    artist: 'Ed Sheeran',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Elton John - I am Still Standing.mid',
-    name: 'I am Still Standing',
-    artist: 'Elton John',
-    difficulty: 'Easy',
-  },
-  {
-    file: "Elvis Presley - Can't Help Falling in Love with You - EASY.mid",
-    name: "Can't Help Falling in Love with You",
-    artist: 'Elvis Presley',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Evanescence - My Immortal.mid',
-    name: 'My Immortal',
-    artist: 'Evanescence',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Game Of Thrones Theme - Easy.mid',
-    name: 'Game Of Thrones Theme Song',
-    artist: 'Ramin Djawadi',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Gravity Falls - Theme.mid',
-    name: 'Gravity Falls Theme Song',
-    artist: 'Brad Breeck',
-    difficulty: 'Easy',
-  },
-  {
-    file: "GunsNRoses - Sweet Child O' Mine.mid",
-    name: "Sweet Child O' Mine",
-    artist: "Guns N' Roses",
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Happy Birthday - Chopin and Liszt style.mid',
-    name: 'Happy Birthday',
-    artist: 'Chopin and Liszt style',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Imagine Dragons - Believer.mid',
-    name: 'Believer',
-    artist: 'Imagine Dragons',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'James Bay - Us.mid',
-    name: 'Us',
-    artist: 'James Bay',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'James Blunt - You Are Beautiful - EASY.mid',
-    name: 'You Are Beautiful',
-    artist: 'James Blunt',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Johann Pachelbel - Canon in D - EASY.mid',
-    name: 'Canon in D',
-    artist: 'Johann Pachelbel',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'John Denver - Take Me Home, Country Roads - EASY.mid',
-    name: 'Take Me Home, Country Roads',
-    artist: 'John Denver',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Justin Bieber - Purpose.mid',
-    name: 'Purpose',
-    artist: 'Justin Bieber',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Linkin Park - In The End (Mellen Gi _ Tommee Profitt Remix).mid',
-    name: 'Linkin Park - In The End (Tommee Profitt Remix)',
-    artist: 'Linkin Park',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Ludwig Van Beethoven - 7th Symphony - 2nd movement - Easy.mid',
-    name: '7th Symphony - 2nd movement',
-    artist: 'Ludwig Van Beethoven',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Maroon 5 - Memories.mid',
-    name: 'Memories',
-    artist: 'Maroon 5',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Maroon 5 - Misery.mid',
-    name: 'Misery',
-    artist: 'Maroon 5',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Metallica_-_Nothing_Else_Matters_piano_solo.mid',
-    name: 'Nothing Else Matters ',
-    artist: 'Metallica',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Mozart - Lacrimosa (Requiem).mid',
-    name: 'Lacrimosa (Requiem).mid',
-    artist: 'Mozart',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Mozart - Sonata In C Major - K545 - EASY.mid',
-    name: 'Sonata In C Major',
-    artist: 'Mozart',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Mozart - Sonata no. 12, K332.mid',
-    name: 'Sonata no. 12',
-    artist: 'Mozart',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Mozart - Turkish March.midi',
-    name: 'Turkish March',
-    artist: 'Mozart',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'My Chemical Romance - Teenagers.mid',
-    name: 'Teenagers',
-    artist: 'My Chemical Romance',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Panic! at the Disco - I Write Sins Not Tragedies.mid',
-    name: 'I Write Sins Not Tragedies',
-    artist: 'Panic! at the Disco',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Passenger - Let Her Go - EASY.mid',
-    name: 'Let Her Go',
-    artist: 'Passenger',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Persona 5 - Hymn of the Soul.mid',
-    name: 'Persona 5 - Hymn of the Soul',
-    artist: '',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Piano Hero 002 - Deadmau5 - Raise Your Weapon (Madeon Remix).mid',
-    name: 'Raise Your Weapon (Madeon Remix)',
-    artist: 'Deadmau5',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Piano Hero 003 - HearthStone - Main Theme.mid',
-    name: 'HearthStone - Main Theme.mid',
-    artist: '',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Piano Hero 007 - Hardwell - Mad World.mid',
-    name: 'Mad World',
-    artist: 'Hardwell',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Piano Hero 009 - Adele - Hello.mid',
-    name: 'Hello',
-    artist: 'Adele',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Piano Hero 010 - DragonForce - Through The Fire And Flames.mid',
-    name: 'Through The Fire And Flames',
-    artist: 'DragonForce',
-    difficulty: 'Hard',
-  },
-  {
-    file: 'Piano Hero 014 - Feed Me - One Click Headshot.mid',
-    name: 'One Click Headshot',
-    artist: 'Feed Me',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Piano Hero 015 - Fort Minor - Remember The Name.mid',
-    name: 'Remember The Name',
-    artist: 'Fort Minor',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Piano_Man_Piano.mid',
-    name: 'Piano Man',
-    artist: 'Billy Joel',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Pokemon_Theme_Song_piano.mid',
-    name: 'Pokemon Theme Song',
-    artist: 'John Siegler & Tamara Loeffler',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Requiem for a dream.mid',
-    name: 'Requiem for a dream',
-    artist: 'Clint Mansell',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'RiverFlowsInYou.mid',
-    name: 'River Flows In You',
-    artist: 'Yiruma',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Sam Smith - Lay Me Down - EASY.mid',
-    name: 'Lay Me Down',
-    artist: 'Sam Smith',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Scott-Joplin-The-Entertainer.mid',
-    name: 'The Entertainer',
-    artist: 'Scott Joplin',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Shawn Mendes - Life of The Party.mid',
-    name: 'Life of The Party',
-    artist: 'Shawn Mendes',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Shawn Mendes, Camila Cabello - Senorita.mid',
-    name: 'Senorita',
-    artist: 'Shawn Mendes, Camila Cabello',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Steven Universe - Amalgam.mid',
-    name: 'Amalgam',
-    artist: 'Steven Universe',
-    difficulty: 'Medium',
-  },
-  {
-    file: 'Steven Universe - Full Theme.mid',
-    name: 'Steven Universe Full Theme',
-    artist: '',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Steven Universe - Love Like You.mid',
-    name: 'Love Like You',
-    artist: 'Steven Universe',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Sting - Every Breath You Take.mid',
-    name: 'Every Breath You Take',
-    artist: 'Sting',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Sting - Fields of Gold.mid',
-    name: 'Fields of Gold',
-    artist: 'Sting',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Stranger Things - Theme.mid',
-    name: 'Stranger Things - Theme.mid',
-    artist: 'Kyle Dixon & Michael Stein',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'The Beatles (John Lennon) - Real Love.mid',
-    name: 'Real Love',
-    artist: 'The Beatles',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'The Beatles (John Lennon) - Twist and Shout.mid',
-    name: 'Twist and Shout',
-    artist: 'The Beatles',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'The Beatles - Let it Be.mid',
-    name: 'Let it Be',
-    artist: 'The Beatles',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'The Cranberries - ZOMBIE.mid',
-    name: 'ZOMBIE',
-    artist: 'The Cranberries',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'The Office.xml',
-    name: 'The Office Theme Song',
-    artist: 'Jay Ferguson',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'The Weeknd - Blinding Lights.mid',
-    name: 'Blinding Lights',
-    artist: 'The Weeknd',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'The_Scientist_-_Coldplay_Piano_Arrangement.mid',
-    name: 'The Scientist',
-    artist: 'Coldplay',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Tones and I - Dance Monkey.mid',
-    name: 'Dance Monkey',
-    artist: 'Tones and I',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Twenty One Pilots - Trees - EASY.mid',
-    name: 'Trees',
-    artist: 'Twenty One Pilots',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Westworld_Theme.mid',
-    name: 'Westworld Theme.mid',
-    artist: 'Ramin Djawadi',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'XXXTENTACION - Sad!.mid',
-    name: 'Sad!',
-    artist: 'XXXTENTACION',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Yiruma - Spring Time.mid',
-    name: 'Spring Time',
-    artist: 'Yiruma',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'canond-easy.xml',
-    name: 'Canon in D Major',
-    artist: 'Pachelbel',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'drake-passionfruit.xml',
-    name: 'Passionfruit',
-    artist: 'Drake',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'fireflies.mid',
-    name: 'Fireflies',
-    artist: 'Owl City',
-    difficulty: 'Medium',
-  },
-  {
-    file: 'lose-yourself.xml',
-    name: 'Lose Yourself',
-    artist: 'Eminem',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'moonlight-sonata.xml',
-    name: 'Moonlight Sonata',
-    artist: 'Beethoven',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'xx-intro.mid',
-    name: 'Intro',
-    artist: 'The XX',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'imagine_dragons-radioactive_piano_version.mid',
-    name: 'Radioactive (new)',
-    artist: 'Imagine Dragons',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Pans_Labyrinth_Lullaby.mid',
-    name: 'Pans Labrynth Lullaby',
-    artist: 'Javier Navarrete',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'All_Eyes_On_Me__Bo_Burnham_Easy_Piano.mid',
-    name: 'All Eyes On Me',
-    artist: 'Bo Burnham',
-    difficulty: 'Medium',
-  },
-  {
-    file: 'Sweet Hibiscus Tea.mid',
-    name: 'Sweet Hibiscus Tea',
-    artist: 'Penelope Scott',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Bloom.Sweet_Hibiscus_Tea Piano Beginner.mid',
-    name: 'Bloom: Sweet Hibiscus Tea',
-    artist: 'Penelope Scott',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Bloom.Let_It_Be Piano Beginner.mid',
-    name: 'Bloom: Let It Be',
-    artist: 'Beatles',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Bloom.Imagine John Lennon Beginner.mid',
-    name: 'Bloom: Imagine',
-    artist: 'Beatles',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Bloom.Into_The_Unknown.mid',
-    name: 'Bloom: Into The Unknown',
-    artist: 'Over The Garden Wall',
-    difficulty: 'Medium',
-  },
-  {
-    file: 'Bloom.Here_Comes_The_Sun The Beatles Piano Beginner.mid',
-    name: 'Bloom: Here Comes The Sun',
-    artist: 'Beatles',
-    difficulty: 'Easy',
-  },
-  {
-    file: "Bloom.People_Can't_Stop_Chillin.mid",
-    name: "Bloom: People Can't Stop Chilling",
-    artist: 'Sports',
-    difficulty: 'Easy',
-  },
-  {
-    file: "Bloom.People_Can't_Stop_Chillin_Added.mid",
-    name: "Bloom: People Can't Stop Chilling With Accompaniments",
-    artist: 'Sports',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Kingdom Hearts - Simple and Clean (antonlab).mid',
-    name: 'Simple and Clean',
-    artist: 'Kingdom Hearts (antonlab)',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Aphex_Twin_Avril_14th.mid',
-    name: 'Avril 14',
-    artist: 'Aphex Twin',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Bloom.Greensleeves.Beginner.mid',
-    name: 'Bloom: Greensleeves',
-    artist: 'Celtic Ladies',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Bloom.Avril14.mid',
-    name: 'Bloom: Avril 14',
-    artist: 'Aphex Twin',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Scarborough_Fair_Easy.mid',
-    name: 'Scarborough Fair',
-    artist: 'Simon and Garfunkel',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Bloom.j1.mid',
-    name: 'j1',
-    artist: 'Jake Fried',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'One_Final_Effort.mid',
-    name: 'Halo 3: One Final Effort',
-    artist: "Martin O' Donnel",
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Gymnopedie_No_1.mid',
-    name: 'Gymnopedie No 1',
-    artist: 'Satie',
-    difficulty: 'Easy',
-  },
-  {
-    file: 'Farruko - Pepas.mid',
-    name: 'Pepas',
-    artist: 'Farruko',
-    difficulty: 'Easy',
-  },
-].map((elem) => ({ ...elem, type: 'song', file: 'music/songs/' + elem.file }))
+const synthesiaManifest = require('./synthesia-manifest.json')
+
+const songs: MusicFile[] = synthesiaManifest.map((s: any) => {
+  return {
+    file: `music/songs/${s.title}.${s.id}.mid`,
+    title: s.title,
+    artist: s.artist ?? s.arranger,
+    difficulty: s.difficulty,
+    source: 'builtin',
+  }
+})
 
 const musicFiles: MusicFile[] = songs
 
