@@ -3,22 +3,17 @@ import type { Song, SongMeasure, SongNote } from '@/types'
 import { getUploadedSong } from '@/features/persist'
 
 /*
- * In development, parse on client.
- * In production, use preparsed songs.
+ * Retrieves
  */
-async function getServerSong(url: string): Promise<Song> {
-  if (url.includes('.xml')) {
-    const xml = await (await fetch('/' + url)).text()
-    return parseMusicXml(xml) as Song
-  }
-  const buffer = await (await fetch('/' + url)).arrayBuffer()
+async function getServerSong(source: string, id: string): Promise<Song> {
+  const buffer = await (await fetch(`/api/midi?source=${source}&id=${id}`)).arrayBuffer()
   return parseMidi(buffer) as Song
 }
 
-export async function getSong(url: string): Promise<Song> {
-  let song = getUploadedSong(url)
+export async function getSong(source: string, id: string): Promise<Song> {
+  let song = getUploadedSong('TODO')
   if (!song) {
-    song = await getServerSong(url)
+    song = await getServerSong(source, id)
   }
   song.notes = song.items.filter((i) => i.type === 'note') as SongNote[]
   song.measures = song.items.filter((i) => i.type === 'measure') as SongMeasure[]
