@@ -1,4 +1,4 @@
-import React, { useEffect, useRef } from 'react'
+import React, { useRef } from 'react'
 import { Toggle, Sizer } from '@/components'
 import { Song, SongConfig, VisualizationMode } from '@/types'
 import { palette as colors } from '@/styles/common'
@@ -18,7 +18,15 @@ export default function SettingsSidebar(props: SidebarProps) {
   const { left, right, visualization, waiting, noteLetter, keySignature } = props.config
   const sidebarRef = useRef<HTMLDivElement>(null)
 
-  useWhenClickedOutside(() => props.onClose?.(), sidebarRef, [])
+  useWhenClickedOutside(
+    () => {
+      if (props.open) {
+        props.onClose?.()
+      }
+    },
+    sidebarRef,
+    [props.open, props.onClose],
+  )
 
   const handleHand = (selected: 'left' | 'right') => {
     if (selected === 'left') {
