@@ -1,4 +1,5 @@
 import type { Song, SongConfig } from '@/types'
+import type { LibrarySong } from '../pages/SelectSong/types'
 import { LOCAL_STORAGE_SONG_LIST_KEY, LOCAL_STORAGE_SONG_SUFFIX } from './constants'
 import Storage from './storage'
 import { UploadedSong } from './types'
@@ -9,9 +10,9 @@ function getStorageFilename(name: string, artist: string) {
 export function getSongStorageKey(name: string, artist: string): string {
   return getStorageFilename(name, artist) + '/' + LOCAL_STORAGE_SONG_SUFFIX
 }
-export function isKeyAlreadyUsed(name: string, artist: string): boolean {
+export function isKeyAlreadyUsed(title: string, artist: string): boolean {
   const songs = getUploadedLibrary()
-  return !!songs.find((s) => s.name === name && s.artist === artist)
+  return !!songs.find((s) => s.title === title && s.artist === artist)
 }
 
 export function getUploadedSong(file: string): Song | null {
@@ -19,25 +20,26 @@ export function getUploadedSong(file: string): Song | null {
   return Storage.get<Song>(storageKey)
 }
 
-export function getUploadedLibrary(): UploadedSong[] {
+export function getUploadedLibrary(): LibrarySong[] {
   if (!Storage.has(LOCAL_STORAGE_SONG_LIST_KEY)) {
-    Storage.set<UploadedSong[]>(LOCAL_STORAGE_SONG_LIST_KEY, [])
+    Storage.set<LibrarySong[]>(LOCAL_STORAGE_SONG_LIST_KEY, [])
   }
-  return Storage.get<UploadedSong[]>(LOCAL_STORAGE_SONG_LIST_KEY) ?? []
+  return Storage.get<LibrarySong[]>(LOCAL_STORAGE_SONG_LIST_KEY) ?? []
 }
 
 /**
  * Need to update song index, as well as individual song data.
  */
-export function saveSong(song: Song, name: string, artist: string): UploadedSong {
-  const songKey = getSongStorageKey(name, artist)
-  const uploadedSong: UploadedSong = {
-    name,
+export function saveSong(song: Song, title: string, artist: string): LibrarySong {
+  const songKey = getSongStorageKey(title, artist)
+  const uploadedSong: LibrarySong = {
+    id: 'TODO', // use md5
+    title,
     artist,
     duration: song.duration,
-    file: getStorageFilename(name, artist),
-    type: 'upload',
-    difficulty: 'N/A,',
+    file: getStorageFilename(title, artist),
+    source: 'upload',
+    difficulty: 0,
   }
   const songs = getUploadedLibrary().concat(uploadedSong)
   Storage.set(songKey, JSON.stringify(song))
