@@ -1,8 +1,8 @@
 export type UploadedSong = {
-  name: string
+  title: string
   artist: string
   duration: number
   file: string
   type: 'upload'
-  difficulty: 'N/A,'
+  difficulty: 'N/A'
 }
